Move check_certificate_input to sui-transaction-checks
