+++ crates/sui-transaction-checks/Cargo.toml
@@ -7,6 +7,7 @@ publish = false
 edition = "2021"
 
 [dependencies]
+once_cell.workspace = true
 sui-macros.workspace = true
 sui-config.workspace = true
 sui-protocol-config.workspace = true
