+++ crates/sui-transaction-checks/src/lib.rs
@@ -7,18 +7,21 @@ pub use checked::*;
 
 #[sui_macros::with_checked_arithmetic]
 mod checked {
-    use std::collections::{BTreeMap, HashSet};
+    use once_cell::sync::OnceCell;
+    use std::collections::{BTreeMap, HashMap, HashSet};
     use std::sync::Arc;
     use sui_config::transaction_deny_config::TransactionDenyConfig;
     use sui_protocol_config::ProtocolConfig;
-    use sui_types::base_types::ObjectRef;
+    use sui_types::base_types::{ObjectID, ObjectRef};
     use sui_types::committee::EpochId;
+    use sui_types::digests::TransactionDigest;
     use sui_types::error::{UserInputError, UserInputResult};
+    use sui_types::executable_transaction::VerifiedExecutableTransaction;
     use sui_types::metrics::BytecodeVerifierMetrics;
     use sui_types::signature::GenericSignature;
-    use sui_types::storage::BackingPackageStore;
     use sui_types::storage::ObjectStore;
     use sui_types::storage::ReceivedMarkerQuery;
+    use sui_types::storage::{BackingPackageStore, GetSharedLocks};
     use sui_types::transaction::{
         InputObjectKind, InputObjects, TransactionData, TransactionDataAPI, TransactionKind,
         VersionedProtocolMessage,
@@ -35,6 +38,7 @@ mod checked {
     };
     use tracing::error;
     use tracing::instrument;
+
     // Entry point for all checks related to gas.
     // Called on both signing and execution.
     // On success the gas part of the transaction (gas data and gas coins)
@@ -138,6 +142,49 @@ mod checked {
             protocol_config,
             epoch_id,
         )?;
+        Ok((gas_status, input_objects))
+    }
+
+    pub fn check_certificate_input<S: ObjectStore, G: GetSharedLocks>(
+        store: &S,
+        shared_lock_store: &G,
+        cert: &VerifiedExecutableTransaction,
+        protocol_config: &ProtocolConfig,
+        reference_gas_price: u64,
+    ) -> SuiResult<(SuiGasStatus, InputObjects)> {
+        // This should not happen - validators should not have signed the txn in the first place.
+        assert!(
+            cert.data()
+                .transaction_data()
+                .check_version_supported(protocol_config)
+                .is_ok(),
+            "Certificate formed with unsupported message version {:?}",
+            cert.message_version(),
+        );
+
+        let tx_data = &cert.data().intent_message().value;
+        let input_object_kinds = tx_data.input_objects()?;
+        let input_object_data = if tx_data.is_end_of_epoch_tx() {
+            // When changing the epoch, we update a the system object, which is shared, without going
+            // through sequencing, so we must bypass the sequence checks here.
+            check_input_objects(store, &input_object_kinds, protocol_config)?
+        } else {
+            check_sequenced_input_objects(
+                store,
+                cert.digest(),
+                &input_object_kinds,
+                shared_lock_store,
+            )?
+        };
+        let gas_status = get_gas_status(
+            &input_object_data,
+            tx_data.gas(),
+            protocol_config,
+            reference_gas_price,
+            tx_data,
+        )?;
+        let input_objects = check_objects(tx_data, input_object_kinds, input_object_data)?;
+        // NB: We do not check receiving objects when executing. Only at signing time do we check.
         Ok((gas_status, input_objects))
     }
 
@@ -334,6 +381,55 @@ mod checked {
                 }
             }
             .ok_or_else(|| SuiError::from(kind.object_not_found_error()))?;
+            result.push(obj);
+        }
+        Ok(result)
+    }
+
+    /// When making changes, please see if get_input_object_keys() above needs
+    /// similar changes as well.
+    ///
+    /// Before this function is invoked, TransactionManager must ensure all depended
+    /// objects are present. Thus any missing object will panic.
+    fn check_sequenced_input_objects<S: ObjectStore, G: GetSharedLocks>(
+        store: &S,
+        digest: &TransactionDigest,
+        objects: &[InputObjectKind],
+        shared_locks_store: &G,
+    ) -> Result<Vec<Object>, SuiError> {
+        let shared_locks_cell: OnceCell<HashMap<_, _>> = OnceCell::new();
+
+        let mut result = Vec::new();
+        for kind in objects {
+            let obj = match kind {
+                InputObjectKind::SharedMoveObject { id, .. } => {
+                    let shared_locks = shared_locks_cell.get_or_try_init(|| {
+                        Ok::<HashMap<ObjectID, SequenceNumber>, SuiError>(
+                            shared_locks_store.get_shared_locks(digest)?.into_iter().collect(),
+                        )
+                    })?;
+                    // If we can't find the locked version, it means
+                    // 1. either we have a bug that skips shared object version assignment
+                    // 2. or we have some DB corruption
+                    let version = shared_locks.get(id).unwrap_or_else(|| {
+                        panic!(
+                            "Shared object locks should have been set. tx_digset: {:?}, obj id: {:?}",
+                            digest, id
+                        )
+                    });
+                    store.get_object_by_key(id, *version)?.unwrap_or_else(|| {
+                        panic!("All dependencies of tx {:?} should have been executed now, but Shared Object id: {}, version: {} is absent", digest, *id, *version);
+                    })
+                }
+                InputObjectKind::MovePackage(id) => store.get_object(id)?.unwrap_or_else(|| {
+                    panic!("All dependencies of tx {:?} should have been executed now, but Move Package id: {} is absent", digest, id);
+                }),
+                InputObjectKind::ImmOrOwnedMoveObject(objref) => {
+                    store.get_object_by_key(&objref.0, objref.1)?.unwrap_or_else(|| {
+                        panic!("All dependencies of tx {:?} should have been executed now, but Immutable or Owned Object id: {}, version: {} is absent", digest, objref.0, objref.1);
+                    })
+                }
+            };
             result.push(obj);
         }
         Ok(result)
